# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API davidWebProject",
        "description": "Cette API est conçue pour gérer l'intégration de diverses fonctionnalités dans mon projet de freelance. Elle permet la gestion complète du backend, incluant la création, la mise à jour, la suppression et la récupération de données essentielles telles que les témoignages clients, les projets, les utilisateurs, et d'autres éléments de contenu. L'API alimente mon site web en fournissant les données nécessaires tout en facilitant la gestion administrative des opérations. Elle est sécurisée via l'authentification par jeton et est optimisée pour une utilisation par des applications web et mobiles. Idéale pour un usage interne, cette API centralise et simplifie la gestion de mon activité freelance.",
        "contact": {
            "name": "Support",
            "url": "https://www.davidwebprojects.fr",
            "email": "contact@davidwebprojects.fr"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://apifreelance.davidwebprojects.fr",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/statuses": {
            "get": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Lister tous les statuts",
                "operationId": "getStatuses",
                "responses": {
                    "200": {
                        "description": "Liste des statuts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Status"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Créer un nouveau statut",
                "operationId": "createStatus",
                "requestBody": {
                    "description": "Données du statut à créer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Status"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Statut créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Status"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/api/statuses/{id}": {
            "get": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Afficher les détails d'un statut spécifique",
                "operationId": "getStatusById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du statut à afficher",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du statut",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Status"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Statut non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Mettre à jour un statut existant",
                "operationId": "updateStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du statut à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Données du statut à mettre à jour",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Status"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statut mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Status"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "404": {
                        "description": "Statut non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Supprimer un statut",
                "operationId": "deleteStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du statut à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Statut supprimé avec succès"
                    },
                    "404": {
                        "description": "Statut non trouvé"
                    }
                }
            }
        },
        "/api/statuses/entity/{entityType}": {
            "get": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Lister les statuts par type d'entité",
                "operationId": "getStatusesByEntityType",
                "parameters": [
                    {
                        "name": "entityType",
                        "in": "path",
                        "description": "Type d'entité pour filtrer les statuts",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des statuts pour le type d'entité spécifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Status"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun statut trouvé pour ce type d'entité"
                    }
                }
            }
        },
        "/proposal-request/temporary-upload": {
            "post": {
                "tags": [
                    "Proposals-Request"
                ],
                "summary": "Uploader un fichier temporaire",
                "operationId": "uploadTemporaryFile",
                "requestBody": {
                    "description": "Fichier à uploader temporairement",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "Le fichier à uploader",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Fichier uploadé temporairement avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "filename": {
                                            "type": "string",
                                            "example": "123e4567-e89b-12d3-a456-426614174000.pdf"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Fichier uploadé temporairement avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation des données échouée"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Status": {
                "title": "Status",
                "description": "Modèle représentant un statut dans le système",
                "required": [
                    "label_status",
                    "entity_type"
                ],
                "properties": {
                    "id": {
                        "description": "ID unique du statut",
                        "type": "integer"
                    },
                    "label_status": {
                        "description": "Libellé du statut",
                        "type": "string"
                    },
                    "entity_type": {
                        "description": "Type d'entité à laquelle le statut est associé (par exemple, 'customer', 'order', etc.)",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Date et heure de création du statut",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date et heure de la dernière mise à jour du statut",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}